# docs/OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GOST-RDPR (Resolve Domains Per Records)",
    "summary": "A utility for working with Mikrotik RouterOS and BGP protocol for announcing IP addresses",
    "description": "The utility provides parsing of domain names into IP addresses, processing of domain lists and their \n    subsequent parsing, processing of individual IP addresses and summarized IP groups. Updates firewall address list \n    and routing table",
    "version": "2.1.0"
  },
  "paths": {
...
    "/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get All Domains Records",
        "description": "Displays all available Domains records",
        "operationId": "Get_all_Domains_records_domains_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "exclusiveMinimum": 0,
              "title": "Limit param",
              "description": "Number of items to be sampled",
              "examples": [
                100
              ],
              "default": 100
            },
            "description": "Number of items to be sampled"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Offset param",
              "description": "Offset quantity to start sampling from",
              "examples": [
                10
              ],
              "default": 0
            },
            "description": "Offset quantity to start sampling from"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 19,
                  "maxLength": 19
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start date",
              "description": "Date from which you want to start sampling. Example `2024-04-24 00:00:00`",
              "examples": [
                "%Y-%m-%d %H:%M:%S",
                "2024-10-01 15:00:00"
              ]
            },
            "description": "Date from which you want to start sampling. Example `2024-04-24 00:00:00`"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 19,
                  "maxLength": 19
                },
                {
                  "type": "null"
                }
              ],
              "title": "End date",
              "description": "Date from which you want to end sampling. Example `2024-04-24 23:59:59`",
              "examples": [
                "%Y-%m-%d %H:%M:%S",
                "2024-10-01 15:00:00"
              ]
            },
            "description": "Date from which you want to end sampling. Example `2024-04-24 23:59:59`"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View resolved domains"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainsPayloadResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Add New Domains",
        "description": "Background add new domains",
        "operationId": "Add_new_domains_domains_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DomainsPostElementReq"
                },
                "examples": [
                  [
                    {
                      "domain": "google.com"
                    },
                    {
                      "domain": "rotterdam1192.discord.gg",
                      "ros_comment": "discord domain"
                    }
                  ]
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataResp"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Clear All Domains Records (Warning!!!)",
        "description": "Clear All Domains records. But not default record id=-1",
        "operationId": "Clear_All_Domains_records__WARNING_____domains_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/domains/search": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Find Domains By Text",
        "description": "Find a Domains by text. Using fields \"name\" or \"url\"",
        "operationId": "Find_Domains_by_text_domains_search_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "exclusiveMinimum": 0,
              "title": "Limit param",
              "description": "Number of items to be sampled",
              "examples": [
                100
              ],
              "default": 100
            },
            "description": "Number of items to be sampled"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Offset param",
              "description": "Offset quantity to start sampling from",
              "examples": [
                10
              ],
              "default": 0
            },
            "description": "Offset quantity to start sampling from"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 19,
                  "maxLength": 19
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start date",
              "description": "Date from which you want to start sampling. Example `2024-04-24 00:00:00`",
              "examples": [
                "%Y-%m-%d %H:%M:%S",
                "2024-10-01 15:00:00"
              ]
            },
            "description": "Date from which you want to start sampling. Example `2024-04-24 00:00:00`"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 19,
                  "maxLength": 19
                },
                {
                  "type": "null"
                }
              ],
              "title": "End date",
              "description": "Date from which you want to end sampling. Example `2024-04-24 23:59:59`",
              "examples": [
                "%Y-%m-%d %H:%M:%S",
                "2024-10-01 15:00:00"
              ]
            },
            "description": "Date from which you want to end sampling. Example `2024-04-24 23:59:59`"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View resolved domains"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "title": "Search text",
              "description": "Search text for fields \"name\""
            },
            "description": "Search text for fields \"name\""
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainsPayloadResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/{id}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get One Domain",
        "description": "Displays one Domain record info",
        "operationId": "Get_one_Domain_domains__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Domain record ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainElementResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResp"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Delete One Domain Record",
        "description": "Background delete one Domain record",
        "operationId": "Delete_one_Domain_record_domains__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": -1,
              "title": "Domain record ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/resolve/check": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Resolve Domain Once",
        "description": "Resolve a domain using configured DNS servers without saving the result to the database",
        "operationId": "Resolve_domain_once_domains_resolve_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainResolveReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResolveResp"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/{id}/resolve": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Resolve Domain Now",
        "operationId": "resolve_domain_now_domains__id__resolve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResp"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DomainElementResp": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "domains_list_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domains List Id"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ros_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ros Comment"
          },
          "ips_v4": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ips V4"
          },
          "ips_v6": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ips V6"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Created At"
          },
          "created_at_hum": {
            "type": "string",
            "title": "Created At Hum"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_at_hum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Hum"
          },
          "last_resolved_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Resolved At"
          },
          "last_resolved_at_hum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Resolved At Hum"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resolved",
          "name",
          "created_at",
          "created_at_hum"
        ],
        "title": "DomainElementResp"
      },
      "DomainResolveReq": {
        "properties": {
          "domain": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 253,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain name",
            "description": "Domain name for one-time resolving without saving to database",
            "examples": [
              "example.com"
            ]
          },
          "id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing domain ID",
            "examples": [
              123
            ]
          }
        },
        "type": "object",
        "title": "DomainResolveReq"
      },
      "DomainResolveResp": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DnsServerResolveResultResp"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "DomainResolveResp"
      },
...
      "DomainsPayloadResp": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "total_resolved": {
            "type": "integer",
            "title": "Total Resolved",
            "default": 0
          },
          "total_query": {
            "type": "integer",
            "title": "Total Query",
            "default": 0
          },
          "payload": {
            "items": {
              "$ref": "#/components/schemas/DomainElementResp"
            },
            "type": "array",
            "title": "Payload",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "limit",
          "offset",
          "duration"
        ],
        "title": "DomainsPayloadResp"
      },
      "DomainsPostElementReq": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain name",
            "examples": [
              "google.com"
            ]
          },
          "list_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domains list id"
          },
          "ros_comment": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Router OS comment for addr-list and route",
            "examples": [
              "discord domain"
            ]
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "DomainsPostElementReq"
      },
      "ErrorResp": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResp"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoDataResp": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "No Data"
          }
        },
        "type": "object",
        "title": "NoDataResp"
      },
      "NotFoundResp": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "default": "NOT_FOUND"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          }
        },
        "type": "object",
        "title": "NotFoundResp"
      },
      "OkResp": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result OK",
            "default": "OK"
          }
        },
        "type": "object",
        "title": "OkResp"
      },
...
    }
  },
...
}